@@ -9,29 +9,31 @@
   "bugs": "https://github.com/dmark/authzero/issues",
   "dependencies": {
     "@oclif/command": "^1.5.19",
-    "@oclif/config": "^1.13.3",
-    "@oclif/plugin-help": "^2.2.1",
-    "@types/dotenv": "^6.1.1",
-    "auth0": "^2.20.0",
+    "@oclif/config": "^1.14.0",
+    "@oclif/plugin-help": "^2.2.3",
+    "auth0": "^2.22.0",
     "dotenv": "^8.2.0",
-    "request": "^2.88.0",
+    "request": "^2.88.2",
     "tslib": "^1.10.0"
   },
   "devDependencies": {
     "@oclif/dev-cli": "^1.22.2",
     "@oclif/test": "^1.2.5",
     "@oclif/tslint": "^3.1.1",
-    "@types/auth0": "^2.9.22",
-    "@types/chai": "^4.2.4",
-    "@types/mocha": "^5.2.7",
-    "@types/node": "^12.11.7",
+    "@types/auth0": "^2.20.0",
+    "@types/chai": "^4.2.9",
+    "@types/mocha": "^7.0.1",
+    "@types/node": "^13.7.4",
+    "@typescript-eslint/eslint-plugin": "^2.20.0",
+    "@typescript-eslint/eslint-plugin-tslint": "^2.20.0",
+    "@typescript-eslint/parser": "^2.20.0",
     "chai": "^4.2.0",
-    "globby": "^10.0.1",
-    "mocha": "^6.2.2",
-    "nyc": "^14.1.1",
-    "ts-node": "^8.4.1",
-    "tslint": "^5.20.0",
-    "typescript": "^3.6.4"
+    "eslint": "^6.8.0",
+    "globby": "^11.0.0",
+    "mocha": "^7.0.1",
+    "nyc": "^15.0.0",
+    "ts-node": "^8.6.2",
+    "typescript": "^3.7.5"
   },
   "engines": {
     "node": ">=8.0.0"
