@@ -1,7 +1,6 @@
 import {Command, flags} from '@oclif/command'
 
-var dotenv = require('dotenv');
-var ManagementClient = require('auth0').ManagementClient;
+import {connectToAuth0} from '../../authzero-helpers'
 
 export default class Client extends Command {
   static description = 'Get client (application) settings.'
@@ -19,14 +18,7 @@ export default class Client extends Command {
     const {args, flags} = this.parse(Client)
     const ID = flags.id
     
-    dotenv.load();
-
-    var auth0 = new ManagementClient({
-      domain: process.env.AUTH0_DOMAIN,
-      clientId: process.env.AUTH0_CLIENT_ID,
-      clientSecret: process.env.AUTH0_CLIENT_SECRET,
-      scope: 'read:users'
-    });
+    const auth0 = connectToAuth0()
 
     auth0.clients.get({ client_id: ID }, function (err: any, client: any) {
       if (err) {
