@@ -1,2 +1,1 @@
-import * as dotenv from 'dotenv'
-dotenv.config()
+require('dotenv').config();
